+++ yaks/admin.py
@@ -1,6 +1,21 @@
+# Copyright (c) 2018 ADLINK Technology Inc.
+#
+# See the NOTICE file(s) distributed with this work for additional
+# information regarding copyright ownership.
+#
+# This program and the accompanying materials are made available under the
+# terms of the Eclipse Public License 2.0 which is available at
+# http://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
+# which is available at https://www.apache.org/licenses/LICENSE-2.0.
+#
+# SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
+#
+# Contributors: Angelo Corsaro, ADLINK Technology Inc. - Yaks API refactoring
+
+
 class Admin(object):
     def __init__(self, ws):
         self.ws = ws
-    
+
     def add_storage(self, selector, properties):
-        return true
+        return True